Dev cloud9

+++ taskboard/app/Http/Controllers/TaskBoardController.php
@@ -34,28 +34,8 @@ class TaskBoardController extends Controller
         {
             $search = $request->search;
         }
-        
-        $taskboardList = $this->taskboardService->getTaskboardList($search);
-
-        // 日時を日付表記に変更
-        $taskboards = array();
-        foreach($taskboardList as $taskboard)
-        {
-            $board = array();
-            
-            $board[] = $taskboard[0];
-            $board[] = $taskboard[1];
-            $board[] = $taskboard[2];
-           
-            // strtotimeを使用してタイムスタンプに変換
-            $timestamp = strtotime($taskboard[3]);
-            
-            // タイムスタンプを年月日の形式で出力
-            $board[] = date("Y年m月d日", $timestamp);
-            
-            $taskboards[] = $board;
-            
-        }
+        // タスクボードの一覧取得        
+        $taskboards = $this->taskboardService->getTaskboardList($search);
 
         return view('taskboards.index', compact('taskboards'));
     }
@@ -92,6 +72,7 @@ class TaskBoardController extends Controller
         // タスクボードの利用者作成
         $taskboardUsers = $this->taskboardService->createOrUpdateTaskboardUsers($taskboardId, $users);
        
+        // 登録完了ページの表示
         $data = [
             'message' => '登録完了しました。',
             'link' => 'taskboards.index',
@@ -106,6 +87,11 @@ class TaskBoardController extends Controller
      */
     public function edit($id)
     {
+        // ログインしているユーザーがタスクボードの利用者でない場合または管理者でない場合
+        if(!$this->taskboardService->judgeLoginUserTaskboard($id) && !$this->userService->judgeUserAdmin()) {
+            return $this->hidden();
+        }
+        
         // タスクボード情報を取得
         $board = $this->taskboardService->getTaskboard($id);
         $taskboard = $board[0];
@@ -127,6 +113,11 @@ class TaskBoardController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // ログインしているユーザーがタスクボードの利用者でない場合または管理者でない場合
+        if(!$this->taskboardService->judgeLoginUserTaskboard($id) && !$this->userService->judgeUserAdmin()) {
+            return $this->hidden();
+        }
+        
         $request->validate([
             'name' => 'required|string|max:255',
             'list' => 'required|array',
@@ -156,6 +147,11 @@ class TaskBoardController extends Controller
      */
     public function destroy($id)
     {
+        // ログインしているユーザーがタスクボードの利用者でない場合または管理者でない場合
+        if(!$this->taskboardService->judgeLoginUserTaskboard($id) && !$this->userService->judgeUserAdmin()) {
+            return $this->hidden();
+        }
+        
         // タスクボードの削除;
         $taskboard = $this->taskboardService->deleteTaskboard([$id]);
 
@@ -173,6 +169,11 @@ class TaskBoardController extends Controller
      */
     public function show($id)
     {    
+        // ログインしているユーザーがタスクボードの利用者でない場合または管理者でない場合
+        if(!$this->taskboardService->judgeLoginUserTaskboard($id) && !$this->userService->judgeUserAdmin()) {
+            return $this->hidden();
+        }
+        
         // タスクボード情報を取得
         $board = $this->taskboardService->getTaskboard($id);
         $taskboard = $board[0];
@@ -188,9 +189,20 @@ class TaskBoardController extends Controller
         // タスクを取得
         $tasks = $this->taskboardService->getTaskboardTasks($id);
 
-        // ユーザー一覧を取得
-        $users = $this->userService->getUserList('');
+        return view('taskboards.taskboard', compact('taskboard', 'userArray', 'userName', 'tasks'));
+    }
+    
+    /*
+    * 禁止ページの表示
+    */
+    public function hidden()
+    {
+        $data = [
+            'message' => 'このページは許可されていないページです。',
+            'link' => 'taskboards.index',
+            'button' => 'タスクボードページ'
+        ];
 
-        return view('taskboards.taskboard', compact('taskboard', 'userList', 'userName', 'tasks', 'users'));
-    } 
+        return view('base.complete', $data);
+    }
 }
